Made StudentMainPage and Shortlist Page connect well to dummy DB.
Shortlist has View bursary option that is able to now take you to view that bursary in another page. Took away the dismiss option and delete bursary from student main page and shortlist page. Created alert to notify if a certain functionality is coming in the next version of app.